Fix stride and slice-height output in some color format.
Particularly, QOMX_COLOR_FORMATYUV420PackedSemiPlanar32m and
QOMX_COLOR_FORMATYUV420PackedSemiPlanar32mMultiView. This is a wild
guess for what stride and slice-height should be in these format
as it is 0 otherwise. I don't know why it's like that, and I don't
know why the playback still works on Android despite that bug.
But for now, this make playback works on Ubuntu Touch.